# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UsuarioRequestDto": {
				"properties": {
					"idPessoa": {
						"type": "number",
						"format": "double"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": [
					"idPessoa",
					"senha"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BasicResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"object": {}
				},
				"required": [
					"message",
					"object"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsuarioDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"idPessoa": {
						"type": "number",
						"format": "double"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"idPessoa",
					"senha"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
		"/usuario": {
			"post": {
				"operationId": "CadastrarUsuario",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UsuarioRequestDto"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "AtualizarSenha",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UsuarioDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletarUsuario",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UsuarioDto"
							}
						}
					}
				}
			}
		},
		"/usuario/id/{id}": {
			"get": {
				"operationId": "FiltrarPorId",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/usuario/all": {
			"get": {
				"operationId": "ListarUsuarios",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/pessoa/id/{id}": {
			"get": {
				"operationId": "FiltrarPorId",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Pessoa"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/livro/id/{id}": {
			"get": {
				"operationId": "FiltrarPorId",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Livro"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/emprestimo/id/{id}": {
			"get": {
				"operationId": "FiltrarPorId",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Emprestimo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/categoria/id/{id}": {
			"get": {
				"operationId": "FiltrarPorId",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Categoria"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}